About to totally change the way validation and serialization is done
