+++ src/items/__init__.py
@@ -1,4 +1,4 @@
 from src.items.item import Item
-from src.items.enums import *
-from src.items.bitmasks import *
+from src.items.item_enums import *
+from src.items.item_bitmasks import *
 from src.items.weapon import OneHandedAxe, TwoHandedAxe
